Add a click-based interactive UI + audio and animations.
This was a big commit the game functions were reworked such that there
is now only one function which is playround that plays a round and also
keeps track of score until reaches a certain number (5) then declares
winner. CSS flexbox and animations were used to create the UI while
Javascript was used to control how, in what order and as a response to
which events do some of  the animations and the audio start playing.

P.S. insightful comments are added for the javascript UI and event
listening maneuvers inside of the .js file.